Create checkIfTheNumberIsInList
If we give a list at last of the code and length of the list in 'length' label and target value we have to check in 'target' label, This code will check if the target value there in list or not. If it is in the list the value in 'result' label will change to one after execution. Else the 'result' label remains zero itself.