Create IsGivenListIsPalindrome
If we give a list at last of the code and length of the list in 'length' label, This code will check if the list is palindrome or not. If it is palindrome, the value in 'isPalindrome' label will change to two after execution. Else the 'isPalindrome' value remains one.